Explicitly instantiate normal
