+++ source/framework/algorithm/source/miscellaneous_operation/normal.cpp.in
@@ -1,0 +1,12 @@
+#include "lue/framework/algorithm/default_policies/normal.hpp"
+#include "lue/framework/algorithm/definition/normal.hpp"
+#include "lue/framework/algorithm/value_policies/normal.hpp"
+
+
+namespace lue {
+
+    using Shape_ = Shape<Count, 2>;
+
+    LUE_INSTANTIATE_NORMAL(ESC(policy::normal::{{Policies}} < {{Element}} >), Shape_);
+
+}  // namespace lue
