+++ source/framework/python/source/submodule.cpp
@@ -80,7 +80,6 @@ namespace lue::framework {
 
     void bind_array_partition_id(pybind11::module& module);
     void bind_locality_id(pybind11::module& module);
-    void bind_normal(pybind11::module& module);
     void bind_timestamp(pybind11::module& module);
 
     void bind_read_array(pybind11::module& module);
@@ -160,8 +159,6 @@ namespace lue::framework {
         {
             bind_timestamp(submodule);
         }
-
-        bind_normal(submodule);
 
         bind_model(submodule);
         bind_progressor(submodule);
